requirements.txt
requests
pandas
colorama
load_dotenv